fix: Move SessionObserverFactory to RE SDK as only the owner of the remote UI (i.e. the RE SDK) can attach observer factories
